[tx_graph] Enable reading and writing transactions

+++ tx_graph/src/write/new_output.rs
@@ -1,7 +1,7 @@
 pub struct NewOutput {
-  pub spending_transaction_id: usize,
-  pub spending_input_index: usize,
+  pub spending_transaction_id: u64,
+  pub spending_input_index: u32,
   pub value: u64,
   // TODO How to handle outputs with multiple addresses?
-  pub address_id: u32,
+  pub address_id: u64,
 }
